传一下orz

+++ Client/src/main/java/businessLogicService/CostBenefitblService/CostBenefitblService.java
@@ -4,11 +4,11 @@ public interface CostBenefitblService {
 
     public String getDate(String message);
 
-    public int showCost(String date);
+    public double showCost(String date);
 
-    public int showBenefit(String date);
+    public double showBenefit(String date);
 
-    public int showProfit(String date);
+    public double showProfit(String date);
 
     public String getReport(String date);
